AssembleReference,  AnalyzeBAMs, and VCF R script updates, and added ReferenceBased script
AssembleReference now runs bbduk twice, to trim each read end while retaining PE info, and K trims from a provided ISSR motif fasta. AnalyzeBAMs now works properly with the samples file. The VCF script conducts several analysis types, but still has not been configured as part of the pipeline -- variables for paths and file names need to be generalized. ReferenceBased is an alternative to AssembleReference, where a reference genome is provided and the reads are trimmed and the directory structure expected by downstream analyses is setup.